noria-client: Introduce unsupported_set_mode
As discussed in the SET statement design doc[0], there are two *correct*
choices for how we can behave when we receive SET statements that might
change the semantics of queries in a way we don't currently support - we
can return an error, which is the current behavior, or we can put that
particular connection in a state where all subsequent statements are
proxied upstream. There are different scenarios that might necessitate
each behavior - for example, a user testing locally might want to be
notified as loudly as possible that their application has issued an
invalid SET statement and so prefer the error approach, whereas a user
running in production might want to avoid potential unforseen reasons
for crashing their application, and so prefer the permanent-proxy
approach.

To allow the user to select the two behaviors, this commit introduces a
new `unsupported_set_mode` flag to noria-client-adapter, which is a
subset of but distinct from a new `UnsupportedSetMode` enum in
noria_client, the latter of which subsumes the existing hidden
`allow_unsupported_set` flag (it's now just a variant of that enum, at
least inside nori-client::backend). If the client issues an unsupported
SET statement when the UnsupportedSetMode is set to Proxy, that backend
enters a permanent mode where all statements will be proxied upstream.
Since this is similar to the way we handle transactions, both this and
the transaction state have been moved to a `ProxyMode` state machine
enum which is now stored as a field of the Backend itself (rather than
on the upstream, where it was duplicated and not really the right
separation of concerns anyway).

This all comes with an integration test for each of psql and mysql
checking that this works, by making sure that after we issue an
unsupported SET statement we get the correct results from the upstream
for a subsequent query whose semantics would be altered by that SET
statement.

[0]:https://docs.google.com/document/d/1FnOyKYqYxJh6oD2zd8HuFxeoSvZtYXH-

Change-Id: I0005339c0b36e16389cb41af05a1fcbeefbf3cd9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2358
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>