ENG-1249: Add CI Checks for Helm Chart adherence to k8s API schemas
The goal of this CL is to deliver a form of CI schema checks which run
against any customer facing Helm charts contained with this repository.

One solution out there to accomplish this is Kubeval.

  > kubeval is a tool for validating a Kubernetes YAML or JSON
    configuration file. It does so using schemas generated from the
    Kubernetes OpenAPI specification, and therefore can validate
    schemas for multiple versions of Kubernetes.

This CL adds Helm template rendering for each of our values files, and
implements dynamic logic to run `kubeval` against each of those
renditions. This should help ease maintainability as we add more files,
charts, etc. and as Kubernetes APIs change going forward.

The kubeval --strict flag is also added to ensure we never generate
YAML declarations with properties which do not exist in the
OpenAPI specs.

Change-Id: Id434754d827c39b839b170438c1bdd2b038b8c22
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2361
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI