noria-psql: Add more allowed parameters with values
I've manually run through the list of all postgresql parameters, and
narrowed it down to the list of parameters that have the potential to
change the semantics of queries in a way that would matter for us. This
commit adds all of those parameters, and the values that match the
semantics we *do* implement, to the set statement handler in noria-psql.

Change-Id: I3de53b72b71b2d1b04ac8a04e3724589be407689
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2359
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>