noria-psql: Add all remaining any-value parameters
Add *all* of the remaining currently supported postgresql parameters,
which I scraped from [0], to the list of postgresql parameters which we
support when set with any value, because they don't change query
semantics at all. This is a very long list, but I'm keeping it opt-in
rather than opt-out so that we behave correctly if future versions of
postgresql add new parameters that would change query semantics in a way
we don't support

I went through this list manually (yes, the whole list) so it's
possible (likely even) that I've missed something here - going forwards,
we'll likely want to do a more careful audit of these parameters, but
for now at least this gets us a lot closer to just being able to handle
new clients.

[0]: https://pgdash.io/blog/postgres-configuration-cheatsheet.html

Change-Id: I2465d5cdff94ef3287721718b9ea20b6ec77d60a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2360
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>