noria-mysql: Add all remaining any-value parameters
Add *all* of the remaining currently supported mysql parameters, which I
scraped from [0], to a new list of mysql parameters which we support
when set with any value, because they don't change query semantics at
all (mirroring the same list which exists in noria-psql). This is a very
long list, but I'm keeping it opt-in rather than opt-out so that we
behave correctly if future versions of mysql add new parameters
that would change query semantics in a way we don't support

I went through this list manually (yes, the whole list) so it's
possible (likely even) that I've missed something here - going forwards,
we'll likely want to do a more careful audit of these parameters, but
for now at least this gets us a lot closer to just being able to handle
new clients.

[0]https://dev.mysql.com/doc/refman/8.0/en/dynamic-system-variables.html

Change-Id: I86a3b4055f05c298c989b6b015a22de3791dba86
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2363
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>