psql-srv: Implement Flush FrontendMessage
The postgres binary protocol allows clients to send a Flush message,
starting with an 'H' (ascii 72), to request that the server flush its
response buffer before continuing (but otherwise not produce any
response). Up to this point we've not seen any clients that
actually *produce* this mesage, so it's been left unparsed and
unimplemented, but it looks like the node `pg` package does actually
send this message sometimes, so we need to handle it properly.

Change-Id: Iafb324c783c3105ad91fef736247d0e4566b033f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2373
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>