ENG-1251: Integrate latest Grafana dashboards into Helm chart
The purpose of this CL is to copy all of the latest Grafana dashboards
into the dashboards/ directory of the ReadySet Helm chart.

Long term, we can delete these from this directory and during the Helm
chart release process we can copy them in from a single location within
the repo (the source of truth, currently the installer).

Change-Id: I1076744e522252db3d490794ebf788892225c6e5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2375
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>