Update framework testing to display literals
Currently we don't display literals from logs in our framework testing.
Because this feature is just for us, we should display literals so we
can get more meaningful data from test failures. This commits adds a
Dockerfile for both mysql and postgres versions of the adapter, which
are identical to the existing dockerfiles, except that the compile time
feature of display_literals is turned on.

This is, in turn, the Dockerfile that is now pulled in for the framework
testing pipeline to be built.

Change-Id: I64741f0053f72647fe7bfaa814dfd83feb902086
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2382
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>