Error out if wrong MySQL DB used in connection
If a MySQL client attempts to switch databases after connecting (i.e.
with "USE some_db") then ReadySet responds with an error: "Tried to use
database that ReadySet is not replicating from". However, a client can
include a specific database when initially connecting, and in that case
the user does not see any error. Instead, ReadySet ignores whatever
database was specified by the client and silently continues using its
own database.

The reason this happens is because the on_init callback only gets called
from the Init command handler. This change adds an additional call to
on_init to the init function which only gets run if a database is
specified by the connecting client.

Change-Id: Ie67ceef7de2239cd5a1d8022d8117ef50d07f332
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2391
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>