installer: Support using existing external DBs
The installer will now provide the option to not automatically spin
up a new database instance as part of the compose file, but instead
allows specifying a pre-existing upstream's address, username,
and password.

If the user chooses "yes" for the option to automatically create an
upstream database, the behavior is the same as before.

Refs: ENG-1258
Change-Id: I32e461c7cb131b6ee71d4c57d00d2ba18f858cfb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2395
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>