Support numeric field references in ORDER BY + GROUP BY
MySQL (but not PostgreSQL, as far as I can tell) allows the ORDER BY and
GROUP BY clauses of a query to contain references to fields in the
SELECT list *by index*, represented syntactically by numbers (which we
would previously parse as integer literals). This commit implements
support for those numeric field references, by:

1. Parsing them specially, using a new `FieldReference` enum with
   `Numeric` and `Expression` variants, which is used in both the
   GroupByClause and OrderClause of the query
2. Adding a rewrite pass which converts all numeric field references to
   expressions by looking them up in the select list, *after* the pass
   that expands stars into explicit lists of columns (this matches
   mysql's behavior)

We briefly considered special-casing
Expression::Literal(Literal::Int(_)) here instead, but doing it this way
felt a lot cleaner and more explicit, at the expense of just a little
bit of extra verbosity.

This all comes with a new logictest which exercises some of this
behavior.

Fixes: ENG-1317
Change-Id: I25feb430edceb6fc207a9f9f67f1dcb964aa9232
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2401
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>