ci: Don't fail on the second build of a commit
The `aws ecr put-image` operation, despite its naming making it seem
idempotent, actually fails with an error if the given image tag already
exists (grr). This commit makes the build-image script, which was
updated in b32ac1f27 (ci: Don't rebuild images if dependent files are
unchanged, 2022-01-11) to use this operation to avoid rebuilding images,
tolerant to this failure by just adding a `|| true` to the command -
getting this to *specifically* handle the failure is sufficiently
annoying, and if the operation fails for a different reason the rest of
the build will just fail anyway because the image won't exist.

Change-Id: If3f198496641e5d74621df96f314f797468d5b9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2407
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>