telemetry-reporter: Init new crate for reporting telemetry
Add a new `readyset-telemetry-reporter` crate, which provides a simple
reusable mechanism for reporting arbitrary telemetry payloads to the
telemetry ingress application, authenticated via a configurable API key.

Since we don't actually have the infrastructure for *issuing* those API
keys spun up yet, this also takes over the hardcoded API key previously
defined in readyset_installer, renaming it to `HARDCODED_API_KEY` for
clarity and to differentiate it from the non-hardcoded API keys we'll
eventually be issuing to users. As a stop-gap solution, we then
special-case this hardcoded API key such that if it's used, no key
validation is done and no telemetry payloads are sent.

Refs: ENG-1323
Change-Id: If5e2f6d0ebc90486d0ceb307a1da32e20fc69d4f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2421
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>