installer: Allow using Auth0 tokens as API tokens
Use the new TelemetryReporter struct in the installer to allow using
auth0-provided JWTs in addition to (for now) the hardcoded API token, in
the installer. Currently this TelemetryReporter is stored as an unused
field on the Installer struct itself, but next up we'll use it to
actually report some telemetry!

Fixes: ENG-1323
Change-Id: I9ebc73f2322f2f01f77ed102393a379ea6b01622
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2425
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>