docs: Update and add a variety of info related to Gerrit
The existing docs link to a Google document that includes a variety of
info on Gerrit, including social norms and conventions along with
defining some terms. This change incorporates that information more
seamlessly into the flow of the existing documentation, and also adds
some additional information for understanding and using Gerrit in general.

Change-Id: I06528378e2528c35b3be84cf34da7ef28c871d4b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2435
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI