Switch framework tests to use explicit migrations
This switches framework tests to use explicit migrations feature. The
framework tests are here mostly so we can understand if users can get
basic connectivity with their framework or client library of choice. To
that end, our most common configuration right now for a customer is for
them to use the explicit migration feature, where they choose which
queries to cache, or not cache.

Because this is our default for how customers will use ReadySet, it is a
better model of an out of the box experience. If we still experience
issues with our framework tests after this change, those should be
pretty large red flags for us to fix quickly.

Change-Id: I09781b4a19433af45ea25a1406d4538b1bf40719
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2437
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>