Fix python/pymysql framework test
The framework test was continually failing to find cryptography, even
though we were compiling it in the docker container. We even would exec
into the container, run the test file from there, and it would run fine,
but wouldn't be able to find the cryptography dep if ran from
docker-compose or docker run. Eventually, I figured out that if you
upgrade pip, there is a muslwheel that supplies cryptography for you.
This method of getting cryptography yields a container that does *not*
complain that it can't find cryptography. Yay!

This commit also refactors the Dockerfile for best python practices by
using venv.

Change-Id: I0c4db631e4806630965cf3f25c64e5b328645f1d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2461
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI