Fix sqlalchemy framework test runtime error
Apparently we need to clone into the workdir directly, and then move our
stuff into that same dir so we are working from some mega directory
before we pip install the sqlalchemy repo.

This also hard codes us to the most recent commit that updated the
examples. It appears there is a real incompatibility if we don't do
this, as some examples are out of date.

Python is fun.

Change-Id: Ifc19bec36b0a4d5fc6cd508131b1c06b36a07e94
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2462
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>