installer: Support in-place upgrade of deployments
If the installer is run with an existing docker-compose deployment, give
the user the option to *upgrade* that deployment in-place. Then, compare
the docker-compose file with the version that the installer *would*
create if run from scratch, and if different (eg because the installer
has been compiled with different docker image tags) download and load
new docker images, replace the template in-place, and run
`docker-compose up -d` on the compose file to cause docker-compose to
recreate the running containers with new images.

Since in a lot of cases this is the third time we've duplicated some
logic, this commit also includes a fair bit of cleanup, reorganization,
and deduplication of said logic, to make things flow a little bit
cleaner.

There's no automated tests for the installer atm, but I've tested this
manually by replacing the docker image tag in constants.rs and
upgrading, then tearing down, an existing deployment, and everything
worked great.

Fixes: ENG-1306
Change-Id: I803b1aa2f1343c08f40b8585d1b4f583ced38e27
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2467
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>