Replace unbounded-interval-tree with merging-interval-tree
Merging-interval-tree is a simpler and faster implementation of
unbounded-interval-tree.

Internally the new tree stores the endpoints of the intervals in
a BTreeSet structure, and it merges intervals whenever they overlap
each other. This allows it to avoid fragmentation and slow downs
even after extended modification.

Benchmarks suggest that the new tree is over 5x faster on inserts,
over 3x faster on insert/delete and over 2x faster on point lookups.

Change-Id: I01515ff4ac835de806b895a6fd8c077dfbadf8b0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2471
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>