installer: Implement first version of smoke test using rexpect
This adds a test for the installer which directly executes the installer
binary and feeds it commands using the rexpect library.

Currently the test just checks the CLI output of the installer, but
future additions could potentially attempt to connect to ReadySet and/or
the dashboard to try to verify that the installed components are
functioning correctly.

Change-Id: If019245b8380f15d3303371e2222746170358c23
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2475
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Buildkite CI