controller: Generalize out `query_domains` function
The pattern of asynchronously sending out a bunch of domain requests
and executing them concurrently with `buffer_unordered` has been
repeated two times now, and is about to be repeated a third time (to get
state sizes for graphviz), so now seems like a great time to generalize
it out into a function. Originally played around with taking a set of
domain indices and a function to generate the request, but eventually
settled on just taking in an iterator of `(domain, request)` tuples,
which feels a lot cleaner.

Change-Id: Ica16c7dc76802462272e0fc43ff8c719682afc11
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2479
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>