deny.toml: Ignore RUSTSEC-2020-0071
I'm not entirely sure what's going on here, but this ignored advisory,
which was mitigated long ago by the introduction of the
clippy::disallowed_method lint in 40e27917c ([CVE-2020-26235] Remove the
usage of `chrono` unsafe methods., 2021-10-20), had *gone away* at some
point, and subsequently was removed from the configuration by
29133c964 (deny.toml: Temporarily ignore recent advisories, 2022-05-23),
but now it's returned and is causing builds to fail agaian. Regardless,
for the same reason we were able to safely ignore it before, it is now
safe to ignore it again.

Change-Id: I14e5414f9fe80b894e73b7914f90b43b9d4a493c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2481
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI