ops/substrate: Add telemetry-snowflake module
Add a Terraform module for provisioning a Snowflake storage integration,
stage, and external table to query the S3 bucket that the
telemetry-ingress is using for storage. Because of some shortcomings in
the way that S3 storage integrations need to be provisioned for
Snowflake, this needs to be set up in two steps, with some copy-pasting
of values between the steps, to avoid a circular dependency. I've
documented that two-step provisioning process in the README.md of the
module.

Change-Id: Id97da8f1b0d69b49002c1e2614dec9118c346be0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2482
Tested-by: Griffin Smith <griffin@readyset.io>
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>