installer: Don't hard-fail on 5xx-like auth failures
If the API call to authenticate the TelemetryReporter fails in the
installer due to an error that looks like it's our fault (a 5xx, an IO
error, a timeout, etc.) vs an invalid API key, just keep going rather
than hard-failing. We're not adding any sort of monitoring to the
telemetry ingress quite yet, which means it might go down and we
wouldn't know about it, and we'd like people to be able to run the
installer even if that happens.

I'm *reasonably* certain that doing this logic in the installer
itself (vs the telemetry_reporter crate) is the correct separation of
concerns, but it's honestly tough to tell since the installer is
currently the only consumer of the telemetry-reporter crate - it's
entirely possible that once eg readyset-server itself starts sending
telemetry we'll want to move this logic inside the telemetry-reporter
instead.

Change-Id: I3f0801bf78aeffd15c72c5a960796736387dcbd7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2485
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>