ops/images: Add build for auth0-frontend
Add Packer configuration for building an AMI that runs the
auth0-frontend application, which is a small nodejs application to serve
a login page for provisioning new users and giving them API tokens to
use to run the installer. This is set up with nodejs running directly on
the instance as a systemd unit, and an nginx reverse proxy forwarding to
that nodejs process (and redirecting http to https).

Change-Id: I83472a946eaa3195508afd9495e86ebbff524dda
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2494
Tested-by: Buildkite CI
Reviewed-by: Sasha Fox <sasha@readyset.io>