Modify reader_map to store Arc<Vec<_>> with sorted data
Currently the reader_map is heavily optimized towards a very specific
benchmark: fetching short rows from the reader at the cost of a single
hashmap lookup, quickly serializing them on the wire under a so called
ReadGuard and calling it a day.

This optimization assumes that serialization is the only step and is
fast, otherwise the ReadGuard is held for too long, blocking any domain
updates. However that assumption is mostly invalid, first of all, the
serialization step is pretty expensive, and in addition most meaningful
queries will have at least some simple "post lookup" processing applied
to it. The worst offender would be "order by" that sorts the rows on
every read query, the second worst are aggregates, that require another
hashmap lookup for every row in the set. Even the simplest queries will
have at the very least to do some column shuffling and allocation, albeit
that is almost free in comparison.

As for local readers, no serialization is required, but the data has
to be cloned before it leaves the reader, which also incurs a penalty.

This change is the first in a series of changes to address this problem,
and it makes the following assumption (which is the assumption that
underlies readyset in general so it is safe to make): reads are more
frequent than writes, therefore performing some compute on writes saves
a lot of compute on reads.

This change in particular modifies the reader map to store rows of
data in a sorted vector, under an Arc pointer.

Having the data sorted at insertion time means that every row insertion
or deletion will require a binary search to be perfomed on the rows for
a given key, which is a log(n) step, as opposed to having an n*log(n)
sorting on every read or having n hashmap operations for aggregates on
every read. So even if there is a write for every read it is still a net
performance win. In addition results return from readyset now have a
deterministic order, which IMHO is a nicer bahavior than dumping random
rows.

The Arc is there so that the remaining post-lookup processing can take
place outside the ReadGuard, unblocking the domains for further writes,
and for local readers allows to avoid cloning and forwarding results
directly for the reader map.

The drawback is that one can no longer use aliasing for the two copies
of the left_right handle, which means reader map memory usage will
increase to almost 2x in the worst case, although in practice it will
be well below that.

Follow up CLs will break post-lookup into pre-insert and post-lookup
steps, will perform post-lookup outside ReadGuard, and will perform
post-lookup across keys, fixing a huge bug as a result.

Change-Id: I5f995c9301e68e02ef997b442b7551c0898d5fd4
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2576
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>