installer: Don't prompt for an API key if none is provided
The auth page to provision new API keys isn't ready yet, and we'd like
to be able to go live with users and start gathering feedback even if we
don't have the ability to collect telemetry from them. To make the
process as smooth as possible, this commit removes the prompt for API
key entirely if none is provided by the user, instead making the
telemetry reporter just a no-op.

Change-Id: I5af2bf6c1b244f48713db81fe18c628d62161dcb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2500
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>