ops/deploy: Enable logging for installer distribution
Enable "standard" logging for the cloudfront distribution for the
installer, pointing at a new "installer_logs" bucket. We're going to use
this for cheap telemetry on the installer while we're working on
provisioning auth tokens for the new telemetry ingress, and set up
snowflake to query this bucket.

Change-Id: Id49894419a46669fcc4998e5e946cc829523c3c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2503
Tested-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>