Dashboard: Remove connected page & misc copy editing
Removes the ReadySet connection status page. Most of this information is
not likely to be helpful to a user. Users can also determine that
ReadySet is still coming up (and snapshotting) when they attempt to
cache a query. However, we will probably have to make that timeout error
message more informative.

This CL also introduces edits:

Workload Page:
- The two sections are changed to "Proxied Queries" and "Cached Queries"
- The "ReadySet Supported?" field is changed to "Supported By ReadySet"

Specific Query:
- The "Query Running in ReadySet" field is changed to "Query Status" and
the values are changed to "Cached" and "Proxied"

Change-Id: I085095a04aeea2287717c2685665d3de8af723d5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2504
Tested-by: Buildkite CI
Reviewed-by: Jay Bensal <jay@readyset.io>