ops/substrate: Refactor + deploy snowflake to prod
Since we're now going to be using snowflake to query more than one
S3 bucket (in more than one account!), refactor the telemetry-snowflake
module into one that *just* sets up the snowflake storage integrations
and the necessary IAM role to grant it access, taking in the database as
a parameter and providing the name of the created storage integration as
an output to be used in the creation of a stage and external_table in
the calling root module. Then, use this newly refactored module to
deploy a production version of the telemetry database and s3 external
table in the prod account, removing it from the sandbox account along
the way.

Change-Id: I373fb7a57666b96c1484023371c5e93b187965a3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2505
Tested-by: Griffin Smith <griffin@readyset.io>
Terraform-Plan-Review: Griffin Smith <griffin@readyset.io>
Reviewed-by: Harley Klein <harley@readyset.io>