substrate: Speed up Substrate Terraform plans
Previously, we would run Terraform validate and then as a separate CI
step run Terraform plan. This combines this into a single step.

In addition, changes the filter on the list of working directories to
run on to be able to ignore the very large "network" subset as it is
the vast majority of the working directories and it does not get changed
very often. Most of them are the peering workspaces which grows
exponentially with the number of regions.

Finally, removed some old code that used handlebars to template out
Buildkite steps.

Change-Id: I8ef8f9c5a06b96b5c6ecdd154c71f0fa30cd9dfe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2507
Reviewed-by: Griffin Smith <griffin@readyset.io>
Tested-by: Griffin Smith <griffin@readyset.io>