ci: Allow installer release without telemetry ingress release
In Buildkite, the only difference between a block step and an input step
is that a block step creates implicit dependencies in all steps after it
on all steps before it. Because of the way we had our steps ordered, the
usage of a block step rather than an input step for the installer made
it so that implicitly a telemetry-ingress release was required before an
installer release could run. This commit changes the installer block
step to an input step so that we can run installer releases in isolation
of everything else, and sets up all the step dependencies so we still
need to wait on tests to pass before we do. Also, this changes the label
of the step to something a little bit more obvious.

Change-Id: Ie79ea8551977ba2eab070ee6e31d38ad972e7617
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2509
Reviewed-by: Harley Klein <harley@readyset.io>
Tested-by: Buildkite CI