Select: Parse Schema/DB qualifier on nom_sql::Table
Parses the `db/schema`.`table`.`column syntax in Select statements.
Propagates the db/schema qualifier through query_graph and MIR. This is
done by the introduction of a new `Relation` struct.

Relations were previously represented as `SqlIdentifier`s in query_graph
and MIR. `Relation` allows us to associate each relation with a
namespace. `Relation` could be implemented as an alias of `Table` once
the `alias` field is removed.

There may be some places in MIR or query_graph where we want to replace
uses of `Table` with `Relation` (e.g., in join::JoinChain). We also want
to update `MirNode::name` to use `Relation` instead of `SqlIdentifier`.

Because we do not yet support multiple schemas/DBs, the schema field of
`Relation` is largely ignored. This is exploited to allow
HashSet<Relation> lookups with &SqlIdentifier to avoid allocating a
`Relation` when we don't need to. There are still some unfortunate
allocations when we have to convert between `Relation` and `Table`, but
these can be removed once `Table::alias` is removed.

Change-Id: I5e528b4410ee12d6fbe68bd5658af7b79081ac64
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2517
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>