recipe: Move "aliasing query" log to avoid confusion
The `assign_alias` function is called every time we add a new query, but
might not actually add an alias for that query (eg if the query is brand
new) - but the log line was happening indiscriminately, leading to some
confusion. This moves the log line into the branch of the condition
where we're actually adding a new alias to avoid that confusion.

Change-Id: Ia6663cc3dc57fdc315871aaa317be5d09a0d7786
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2524
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>