noria-client: Implement DROP ALL CACHES
Implement the DROP ALL CACHES statement in noria-client, by issung a
`remove_all_queries` RPC to noria_server then clearing all internal
query caches. This includes a slight refactor to some (now more
frequently) repeated patterns within the query status cache code itself.

Given the granularity of the locking in query caches here,
I'm a bit worried about a race condition where a cache gets added in
between the request to remove all queries in noria and the clearing of
the internal list of query caches, but it's also possible that the way
we handle pending queries will also make that totally fine, instead just
requiring a request to noria to fetch the view for that query -
unfortunately this is basically impossible to reliably test, so we'll
mostly just have to watch out for it. The symptom would be a query
incorrectly going to upstream, though, which is far from disastrous, so
I figure this is fine for now, especially since the DROP ALL CACHES
request itself is not likely to be used frequently, or under high-load
scenarios.

Fixes: ENG-1319
Change-Id: Iaee33994c41c1ab4a5837b65e9402b3b0d1d5497
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2525
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>