reader-map: Differentiate between not ready and destroyed
Currently, operations on a reader_map can return None if the map has
never been published yet (which in the case of a dataflow reader can
mean that a full replay has not yet completed, for example), *or* if the
map has been destroyed (which can mean that a reader was dropped from
the graph as part of a migration, eg because a dependent table was
dropped). Because we're unable to distinguish between these two cases,
we always assume from the outside that it's the not-ready case.
The *right* thing to do in the client if we attempt a read in that case
is to wait, since it's safe to assume that the reader eventually will
become ready, but that doesn't work in the dropped case - what we should
do instead is remove that view from the cache and attempt to look it up
again later. This was what was causing ENG-1325, since we always assumed
a dropped reader was just not ready yet.

To allow us to start thinking about differentiating between these two
cases, this commit introduces a reader_map::Error enum with NotPublished
and Destroyed as variants, and returns that wrapping the
previously-present return values in most of the cases where we were just
returning Option. Next up, we'll thread this error type up through
Result return values in reader lookups, then differentiate between the
different types of errors in the client to remove the cached view in the
case that the reader was destroyed.

Refs: ENG-1325
Change-Id: I4f84b2ca9da5e2e59d6c39ac7bbb589cf9ab17a8
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2535
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>