server: Propagate Destroyed error through View RPC
Properly thread through the new Destroyed error that we're now
distinguishing from NotReady for reader maps up through the read
handling code, and return it from the View read RPC as a new
ReadySetError::ViewDestroyed variant. Next, we can use this in the
client to get rid of cached view handles if we try to do lookups on
destroyed views.

Refs: ENG-1325
Change-Id: I55253ebf97f060e8267100860016ed0b42fb5dea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2536
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI