Remove skeema testing from framework pipeline
Currently the skeema test fails, because the way that skeema works, is
that it creates a temporary database, and applies new schema to that
temp database, using that as a method of diffing against the main
database.

We silently allow skeema to think it has connected to a new temp
database it created, even though we are only ever connected to the
primary database. When it tries to issue a CREATE TABLE statement for a
table that already exists in that primary database, we fail the test.

We should inform customers that if they want to use a tool like skeema,
that they should point it directly at their upstream (mysql/postgres)
database. Until we can support connecting to multiple databases, there's
not value in running this test in our pipeline.

Keeping the skeema framework test code itself around in case we want to
run it again in the future when we support multiple databases.

Change-Id: I2a2c75e15ee7ca5cf63e6622450158a5e64aa99e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2540
Tested-by: Buildkite CI
Reviewed-by: Corbin Hall <sif@readyset.io>