dataflow: More consistently use ReplicaAddress
Throughout dataflow and server, we use tuples of `(DomainIndex, usize)`
to refer to a single shard of a single domain. There was a type alias
for this tuple, called `ReplicaAddr`, which was used *a bit* in
dataflow, but not really in enough places - and because it's a type
alias it was too easy to just use the tuple instead, which has all of
the usual downsides of tuples - it's not obvious from the type what that
usize is, it's not obvious from `.0` or `.1` element accesses what
element is being accessed, and modifying the tuple type down the road
requires a significant amount of shotgun-surgery.

This commit introduces a new `ReplicaAddress` struct, which supplants
the old `ReplicaAddr` tuple, and uses it in as many places as I could
find where we were passing around domain index and shard index next to
each other, updating field and other naming to be consistent along the
way.

One notable *exception* to the change here is the
InitialState::PartialGlobal type, which had a `trigger_domain` field
which included not the shard index of the domain, but the *number of
shards in the domain*! Obviously that was super confusing since it's
still just a usize, but now that it's been broken out into a field it
should be a lot more obvious what that means.

Change-Id: Iccf4173c7a51e65a141be6ee2fc3e61d0059d826
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2546
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>